feat: editor de video (local/remoto) + reproductor, badges de próximo a subir, subida TUS resumable
- Nuevo VideoDetailSheet/RemoteVideoDetailSheet + VideoPlayerDialog (local y
  remoto): ver/reproducir y editar links de plataforma directo desde
  Biblioteca, sin pasar por el flujo de Subir.
- TUSUploadClient (core/network) para subidas resumables a Biblioteca remota.
- LibraryScreen y el picker de Subir ahora muestran un badge "Próximo YT/IG/TT"
  sobre el video que el calendario (GET /api/sync/calendar-config) marca como
  el siguiente a publicar en cada plataforma.
- Achica los botones de editar/eliminar en LibraryItemCard (48dp -> 36dp) y
  agrega ellipsis al título -- antes no dejaban espacio para ver el nombre
  completo del archivo.

+++ feature/sync/src/main/kotlin/com/esseanalytics/android/feature/sync/SyncScreen.kt
@@ -430,7 +430,7 @@ private fun ReviewCard(
         Column(Modifier.padding(12.dp)) {
             Text(item.title, style = MaterialTheme.typography.bodyMedium, maxLines = 2)
             Text(
-                formatDuration(item.durationSeconds),
+                formatDuration(item.durationSeconds.toInt()),
                 style = MaterialTheme.typography.labelSmall,
                 color = MaterialTheme.colorScheme.onSurfaceVariant,
             )
@@ -454,7 +454,7 @@ private fun ReviewCard(
                             Column(modifier = Modifier.weight(1f)) {
                                 Text(candidate.file_name, style = MaterialTheme.typography.bodySmall, maxLines = 1)
                                 Text(
-                                    formatDuration(candidate.duracion_segundos),
+                                    formatDuration(candidate.duracion_segundos.toInt()),
                                     style = MaterialTheme.typography.labelSmall,
                                     color = MaterialTheme.colorScheme.onSurfaceVariant,
                                 )
